Implemeted more stuff. Also tests.

+++ src/main/java/ca/jrvs/apps/twitter/dto/Coordinates.java
@@ -2,6 +2,6 @@ package ca.jrvs.apps.twitter.dto;
 
 public class Coordinates {
 
-  public Float[] coordinates;
+  public Double[] coordinates;
   public String type;
 }
